Mostrar resultado

+++ mainengrupo/src/calcularmitjana.java
@@ -1,13 +1,15 @@
 public class calcularmitjana {
+    
     public static double calcularMedia(afegirnota alumno) {
-        int[] notas = alumno.getNumNotas();
-        int total = 0;
+
+        double[] notas = alumno.getNumNotas();
+        double total = 0;
 
         for (int i = 0; i < notas.length; i++) {
             total += notas[i];
         }
 
-        return (double)total/notas.length;
+        return total/(double)notas.length;
     }
 
 }
